Extract Spandx::Cli::Main class
