+++ lib/spandx/cli.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'thor'
+
+module Spandx
+  module Cli
+    Error = Class.new(StandardError)
+  end
+end
