--- lib/spandx/cli/templates/index/update/.gitkeep
@@ -1,1 +1,0 @@
-#
