--- lib/spandx/cli/templates/scan/.gitkeep
@@ -1,1 +1,0 @@
-#
